+++ buildpacks/ruby/src/layers/shared.rs
@@ -76,61 +76,4 @@ mod tests {
         }
     }
     migrate_toml_chain! {TestMetadata}
-
-    #[test]
-    fn test_cached_layer_write_metadata_restored_layer_action() {
-        let temp = tempfile::tempdir().unwrap();
-        let context = temp_build_context::<RubyBuildpack>(temp.path());
-
-        // First write
-        let result = cached_layer_write_metadata(
-            layer_name!("testing"),
-            &context,
-            &TestMetadata {
-                value: "hello".to_string(),
-            },
-        )
-        .unwrap();
-        assert!(matches!(
-            result.state,
-            LayerState::Empty {
-                cause: EmptyLayerCause::NewlyCreated
-            }
-        ));
-
-        // Second write, preserve the contents
-        let result = cached_layer_write_metadata(
-            layer_name!("testing"),
-            &context,
-            &TestMetadata {
-                value: "hello".to_string(),
-            },
-        )
-        .unwrap();
-        let LayerState::Restored { cause } = &result.state else {
-            panic!("Expected restored layer")
-        };
-        assert_eq!(cause.as_ref(), "Using cache");
-
-        // Third write, change the data
-        let result = cached_layer_write_metadata(
-            layer_name!("testing"),
-            &context,
-            &TestMetadata {
-                value: "world".to_string(),
-            },
-        )
-        .unwrap();
-
-        let LayerState::Empty {
-            cause: EmptyLayerCause::RestoredLayerAction { cause },
-        } = &result.state
-        else {
-            panic!("Expected empty layer with restored layer action");
-        };
-        assert_eq!(
-            cause.as_ref(),
-            "Clearing cache due to change: value (hello to world)"
-        );
-    }
 }
